refactor(libs-frontend-shared): #923 clean up crud modal state and added docs
