+++ libs/frontend/shared/src/components/modal/index.ts
@@ -1,2 +1,4 @@
 export * from './CrudModal'
 export * from './crudModalsState'
+export * from './useCrudModalState'
+export * from './useMutationCrudForm'
